Fix issue with login redirect persistence after failed token attempt.
The verify_otp view relies on the get or post param 'next' to store
a redirect url that can override the default settings.LOGIN_REDIRECT_URL.

Upon a failed post it would attempt to pull the next param
from a request GET, which of course there is none.  This caused the
next param to reset to the default settings.LOGIN_REDIRECT_URL after
a failed post.